+++ src/main/java/ru/pethelper/service/PetService.java
@@ -5,9 +5,9 @@ import ru.pethelper.domain.Pet;
 import java.util.List;
 
 public interface PetService {
-    void addPet(Pet pet, int userId);
+    void addPet(Pet pet, long userId);
 
-    List<Pet> getAll(int userId);
+    List<Pet> getAll(long userId);
 
-    Pet getPet(int petId);
+    Pet getPet(long petId);
 }
